e2elive.py can fetch test data from S3 some S3 code moves to util
