+++ misc/util.py
@@ -142,3 +142,20 @@ def ensure_test_db(connection_string, keep_temps=False):
     else:
         logger.info("leaving db %r", dbname)
     return 'dbname={} sslmode=disable'.format(dbname)
+
+# whoever calls this will need to import boto and get the s3 client
+def firstFromS3Prefix(s3, bucket, prefix, desired_filename, outdir=None, outpath=None):
+    response = s3.list_objects_v2(Bucket=bucket, Prefix=prefix, MaxKeys=10)
+    if (not response.get('KeyCount')) or ('Contents' not in response):
+        raise Exception('nothing found in s3://{}/{}'.format(bucket, prefix))
+    for x in response['Contents']:
+        path = x['Key']
+        _, fname = path.rsplit('/', 1)
+        if fname == desired_filename:
+            if outpath is None:
+                if outdir is None:
+                    outdir = '.'
+                outpath = os.path.join(outdir, desired_filename)
+            logger.info('s3://%s/%s -> %s', bucket, x['Key'], outpath)
+            s3.download_file(bucket, x['Key'], outpath)
+            return
